[Updating Role.java]

+++ src/main/java/com/gl/empmgnt/model/Role.java
@@ -18,7 +18,7 @@ public class Role {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO, generator = "native")
 	@GenericGenerator(name = "native", strategy = "native")
-	private int roleId;
+	private long roleId;
 
 	private String roleName;
 }
